GEODE-2420: Enable export logs size estimation and user warning
Adds 'export logs' option, --file-limit-size, to allow user to set
maximun size of the epxorted logs zip file.

When size checking is enabled (file-limit-size > 0) then the check
will also prevent filling up the disk on each member while consolidating
and filtering the logs.

+++ geode-core/src/test/java/org/apache/geode/management/internal/cli/commands/ExportLogsFileSizeLimitTest.java
@@ -77,7 +77,7 @@ public class ExportLogsFileSizeLimitTest {
     assertThat(exportLogsCommand.isFileSizeCheckEnabledAndWithinLimit(0, file)).isFalse();
   }
 
-  private void fillUpFile(File file, int sizeInBytes) throws IOException {
+  private void fillUpFile(File file, long sizeInBytes) throws IOException {
     PrintWriter writer = new PrintWriter(file, "UTF-8");
     while (FileUtils.sizeOf(file) < sizeInBytes) {
       writer.println("this is a line of data in the file");
